refactor: 전역 예외 처리를 위한 패키지 구조 재구성
