--- moamoa/src/main/java/com/likelion/moamoa/domain/auth/exception/NotFoundLoginIdException.java
@@ -1,9 +1,0 @@
-package com.likelion.moamoa.domain.auth.exception;
-
-import com.likelion.moamoa.global.exception.BaseException;
-
-public class NotFoundLoginIdException extends BaseException {
-    public NotFoundLoginIdException() {
-        super(AuthErrorCode.USER_NOT_FOUND_LOGINID_404);
-    }
-}
