+++ moamoa/src/main/java/com/likelion/moamoa/global/response/code/folder/NotFoundUserException.java
@@ -1,6 +1,5 @@
-package com.likelion.moamoa.domain.folder.exception;
+package com.likelion.moamoa.global.response.code.folder;
 
-import com.fasterxml.jackson.databind.ser.Serializers;
 import com.likelion.moamoa.global.exception.BaseException;
 
 public class NotFoundUserException extends BaseException {
