+++ moamoa/src/main/java/com/likelion/moamoa/global/response/code/user/DuplicateEmailException.java
@@ -1,0 +1,9 @@
+package com.likelion.moamoa.global.response.code.user;
+
+import com.likelion.moamoa.global.exception.BaseException;
+
+public class DuplicateEmailException extends BaseException {
+    public DuplicateEmailException() {
+        super(UserErrorCode.USER_DUPLICATE_EMAIL_409);
+    }
+}
