+++ moamoa/src/main/java/com/likelion/moamoa/global/response/code/user/InvalidPasswordException.java
@@ -1,9 +1,9 @@
-package com.likelion.moamoa.domain.auth.exception;
+package com.likelion.moamoa.global.response.code.user;
 
 import com.likelion.moamoa.global.exception.BaseException;
 
 public class InvalidPasswordException extends BaseException {
     public InvalidPasswordException() {
-        super(AuthErrorCode.USER_INVALID_PASSWORD_401);
+        super(UserErrorCode.USER_INVALID_PASSWORD_401);
     }
 }
