+++ moamoa/src/main/java/com/likelion/moamoa/global/response/code/user/UserErrorCode.java
@@ -1,4 +1,4 @@
-package com.likelion.moamoa.domain.auth.exception;
+package com.likelion.moamoa.global.response.code.user;
 
 import com.likelion.moamoa.global.response.code.BaseResponseCode;
 import lombok.AllArgsConstructor;
@@ -6,11 +6,11 @@ import lombok.Getter;
 
 @Getter
 @AllArgsConstructor
-public enum AuthErrorCode implements BaseResponseCode {
+public enum UserErrorCode implements BaseResponseCode {
 
     USER_INVALID_PASSWORD_401("USER_INVALID_PASSWORD_401", 401,"비밀번호가 일치하지 않습니다."),
-    USER_NOT_FOUND_LOGINID_404("USER_NOT_FOUND_LOGINID_404", 404, "해당 ID가 존재하지 않습니다."),
-    USER_DUPLICATE_LOGINID_409("USER_DUPLICATE_LOGINID_409",409,"이미 존재하는 ID입니다.");
+    USER_NOT_FOUND_EMAIL_404("USER_NOT_FOUND_EMAIL_404", 404, "해당 이메일이 존재하지 않습니다."),
+    USER_DUPLICATE_EMAIL_409("USER_DUPLICATE_EMAIL_409",409,"이미 존재하는 이메일입니다.");
 
     private final String code;
     private final int httpStatus;
